Fix git-flow issues with shFlags
I changed the install process to pull the gitflow repo recursively
before running the installer rather than let the installer do it,
because for some reason the installer thinks that the shFlags
submodule is there (when it isn't).

Going by the msysgit-install.cmd I copy the shflags file over the
gitflow-shFlags file which ultimately fixes the bug causing
git-flow not to work on a fresh install.

I also refactored the post_install script so that all the cmd calls
are in the same Invoke-Expression call