Add Paint.NET
Installs & configures paint.net to run in portable mode
https://blog.getpaint.net/2017/07/21/portable-mode-in-paint-net-4-0-17/.

The downloaded file is a .zip file which is extracted to reveal a .exe
file. We then extract the contents of this .exe which gives us a bunch
of files we don't need, plus two 30MB .msi files: PaintDotNet_x64.msi
and PaintDot_x86.msi.

Despite the names, and that these 2 files are slightly different, the
extracted contents for both are exactly the same (apart from one .msi
with the same name as the original except now only 500KB): both 64 and
32-bit libraries are included. Paint.net detects whether to run as 64 or
32-bit on startup. (Though I haven't tested it on a 32-bit system).