Create xwechat.json (wechat4.0)
WeChat 4.0 is a major upgrade of WeChat. Now the official 4.0.3 version has been released.

The original `%AppData%\Tencent\Wechat` changed to `%AppData%\Tencent\xwechat`.
And the original FileSavePathfile changed from `WeChat Files` to `xwechat_files`.

To protect the data integrity of users using WeChat 3.9 and the right to choose the version, instead of forcing them to switch to version 4.0, I named the manifest of version 4.0 `xwechat`.